Refactor gfxdraw
