+++ buildconfig/stubs/mypy_allow_list.txt
@@ -13,8 +13,9 @@ pygame\._sdl2\..*\.__pyx_.*__
 pygame\.examples.*
 pygame\.tests.*
 
-# TODO: Fix error with camera
+# don't look for stubs for modules that are not auto-imported
 pygame\.camera
+pygame\.gfxdraw
 
 # don't look for stubs for pyinstaller hook
 pygame\.__pyinstaller.*
