+++ src/gfxdraw/gfxdraw_doc.h
